+++ src/Core/FrameTimer.cpp
@@ -16,7 +16,7 @@ void FrameTimer<RESULT_IF_INACTIVE>::begin(uint32_t framesToWait_)
 template <bool RESULT_IF_INACTIVE>
 void FrameTimer<RESULT_IF_INACTIVE>::beginAt(uint32_t framesToWait_, float progress_)
 {
-    m_currentFrame = framesToWait_ * progress_;
+    m_currentFrame = static_cast<uint32_t>(framesToWait_ * progress_);
     m_framesToWait = framesToWait_;
 }
